Moved spv bindinginfo and bindingmap
